+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Allow to reconstruct the original text + allow to add a start/end field in as.data.frame (useful but undocumented feature). Set up mainly to be used with the crfsuite R package
 - Added txt_tagsequence
+- Added 1 general function called udpipe which does annotation of data in TIF format.
 
 # CHANGES IN udpipe VERSION 0.6.1